Corrigido arquivo update_p.h com um fdp dum ; que tinha a mais e que demorou 4 dias pra ser achado

+++ trunk/generateutterances1.m
@@ -7,10 +7,10 @@ x1 = [normrnd(1,2,1,nAmostras1);normrnd(5,2,1,nAmostras1)];
 x2 = [normrnd(2,1,1,nAmostras2);normrnd(6,1,1,nAmostras2)];
 
 fid=fopen("loc1.dat","wb");
-fwrite(fid,x1,"double");
+fwrite(fid,x1,"float");
 fclose(fid);
 
 fid=fopen("loc2.dat","wb");
-fwrite(fid,x2,"double");
+fwrite(fid,x2,"float");
 fclose(fid);
